fix(sentry): split brief render module; escaped quotes; repo allowlist
Codex #1769 round 9 — 3 P2s, all peripheral polish (the archive/brief race core
stays converged, nothing re-raised).

1. File-size soft cap. The brief leg reached 678 lines, over the 600-line soft
   cap in docs/pr-checklists/recurring-review-patterns.md (scripts/ has no
   max-lines lint, so the gate did not catch it). Extract the pure rendering,
   the two markdown escapes, and the inert-block assertion into
   scripts/sentry-triage-brief-render.mjs; brief.mjs re-exports them, so no
   importer or test changed. brief.mjs is now 471 lines, the render sibling 231.
   A new test pins both under 600.

2. Escaped quotes in inline lists. splitQuotedListItems (round 8) treated a
   backslash-escaped quote as the end of the scalar, so
   `["Yes: preserve \"A, B\"", ...]` split the inner comma into a corrupted
   bullet. It now honors YAML double-quote backslash escapes and doubled quotes
   of either kind. Test covers both.

3. Displayed repository allowlist. renderBriefComment inlined the raw
   agent-emitted affected_repo into the "How to check — in <repo>" pointer, so a
   prompt-injected attacker/evil-repo could direct a human at an
   attacker-controlled repo (ADR 0036). It now runs through validateAffectedRepo
   (the same allowlist projection uses): an allowlisted external repo or this
   repo renders, anything unrecognized is omitted. Test asserts a hostile repo
   is never rendered verbatim and an allowlisted one still is.

Also routes brief-render.mjs changes to the focused brief suite in
agent-quality-gate.sh.
